Adds new :empty? canned method to double
The SDK has changed and somewhere it calls this method, whereas
previously it didn't.

It is hard to see where this occurs as the Error class it relates to
is dynamically generated.

The rspecs pass when this method is introduced. And given that
the mocked 'response' object has a body it is not 'empty?' and
so I think this is an acceptable fix.