Avoid StackOverflow error when parsing JSON.
On master we get this:

```
response
=> #<struct Aws::CostExplorer::Types::GetDimensionValuesResponse
 dimension_values=[#<struct Aws::CostExplorer::Types::DimensionValuesWithAttributes value="000000000001", attributes={"description"=>"Linked Account name"}>],
 return_size=1,
 total_size=1,
 next_page_token=nil>
```

And

```
 response.to_json
=> "{\"dimension_values\":[{\"value\":\"000000000001\",\"attributes\":{\"description\":\"Linked Account name\"}}],\"return_size\":1,\"total_size\":1,\"next_page_token\":null}"
```

On the update feature branch we get:

```
response
=> #<struct Aws::CostExplorer::Types::GetDimensionValuesResponse
 dimension_values=[#<struct Aws::CostExplorer::Types::DimensionValuesWithAttributes value="000000000001", attributes={"description"=>"Linked Account name"}>],
 return_size=1,
 total_size=1,
 next_page_token=nil>

response.to_json
SystemStackError: stack level too deep
```

There is some kind of recursive feedback loop going on which there are numerous half solutions for online.

But with a call to .to_h first we get:

```
response.to_h.to_json
=> "{\"dimension_values\":[{\"value\":\"000000000001\",\"attributes\":{\"description\":\"Linked Account name\"}}],\"return_size\":1,\"total_size\":1}"
```

Note that next page token is ignored if it us `nil` whereas previously it was converted to a JSON null value.

The tests pass with this formulation but I would like a second opinion on this.